xuan-docx-to-wechat-html/requirements.txt
python-docx>=1.1.0
lxml>=4.9.0
Pillow>=10.0.0
